Build job handler

+++ app/Jobs/ProcessVideo.php
@@ -8,6 +8,7 @@ use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Support\Facades\Storage;
+use Symfony\Component\Process\Process;
 
 class ProcessVideo implements ShouldQueue
 {
@@ -32,7 +33,13 @@ class ProcessVideo implements ShouldQueue
      */
     public function handle()
     {
-        // Do job here
+        $process = new Process("process-v {$this->path}");
+        $process->run();
+
+        if (!$process->isSuccessful()) {
+                throw new ProcessFailedException($process);
+        }
+
         Storage::disk('uploads')->delete($this->path);
     }
 }
